+++ tests/unit/core/test_core.py
@@ -24,6 +24,22 @@ class TestRunway(object):
         assert result.variables == runway_config.variables
         assert result.ctx == runway_context
 
+    def test_init_undetermined_version(self, caplog, monkeypatch,
+                                       runway_config, runway_context):
+        """Test init with unsupported version."""
+        monkeypatch.setattr(MODULE + '.__version__', '0.1.0-dev1')
+        caplog.set_level(logging.WARNING, logger=MODULE)
+        assert Runway(runway_config, runway_context)
+        assert 'shallow clone of the repo' in '\n'.join(caplog.messages)
+
+    def test_init_unsupported_version(self, monkeypatch, runway_config,
+                                      runway_context):
+        """Test init with unsupported version."""
+        monkeypatch.setattr(MODULE + '.__version__', '1.3')
+        with pytest.raises(SystemExit) as excinfo:
+            assert not Runway(runway_config, runway_context)
+        assert excinfo.value.code == 1
+
     @patch(MODULE + '.components.Deployment')
     def test_deploy(self, mock_deployment, runway_config, runway_context):
         """Test deploy."""